Move allow comment functions to new file
